+++ client/components/content-elements/tce-jodit-html/edit/plugins/mdi-icons/index.js
@@ -1,4 +1,3 @@
-import { createButton, createIcon, parent } from './helpers';
 import autoBind from 'auto-bind';
 import { getMdiIcon } from './toolbar-icons';
 
@@ -194,7 +193,9 @@ export default class MdiIconsPlugin {
    * @param {HTMLElement} picker
    */
   onColorChange(e, picker) {
-    const button = parent(e.target, '[data-color]');
+    const { constructor: Jodit } = this.jodit;
+
+    const button = Jodit.modules.Dom.up(e.target, el => el.matches('[data-color]'), picker);
     if (!button) return;
 
     const selected = picker.querySelector('.active');
@@ -290,4 +291,30 @@ function replaceListener(jodit, target, events, listener, oldListener) {
   jodit.events
     .off(target, events, oldListener)
     .on(target, events, listener);
+}
+
+/**
+ * @param {Object} options
+ * @param {String} options.icon
+ * @param {String} options.text
+ * @param {Number} [options.tabIndex=0]
+ * @returns {HTMLSpanElement}
+ */
+export function createButton({ icon, text, tabIndex = 0 }) {
+  const btn = document.createElement('span');
+  btn.tabIndex = tabIndex;
+  btn.setAttribute('role', 'button');
+  btn.appendChild(createIcon(icon));
+  btn.innerHTML += text;
+  return btn;
+}
+
+/**
+ * @param {String} name
+ * @returns {HTMLSpanElement}
+ */
+export function createIcon(name) {
+  const icon = document.createElement('span');
+  icon.classList.add('jodit_icon', 'mdi', name);
+  return icon;
 }
